toasters et bonne routes produit et utilisateur

+++ src/app/products/pages/page-product-list/page-product-list.component.ts
@@ -25,6 +25,7 @@ export class PageProductListComponent implements OnInit {
 
 
   public deleteProduct(productId: number) {
+    if (confirm('Voulez-vous vraiment supprimer ce produit ?'))
     this.productService.deleteItemById(productId).subscribe(
       () => {
         this.status = "Supprimé";
